Split the date transformer into one class per strategy
date_encoding.py held two unrelated implementations behind a boolean:
reading a point in time as nanoseconds since the epoch, and expanding it
into calendar features with skrub. Neither reads the other's code, so
they are now NumericalDateTransformer and SkrubDateTransformer under
preprocessing/datetimes/, with make_date_transformer mapping the
TRANSFORM_DATES flag onto the choice.

The base class keeps only what the choice does not affect: finding the
temporal columns, converting durations to seconds, and the shape of what
fit_transform hands back. Tests move alongside, one module per strategy.

No behaviour change. The warning's stacklevel goes from 6 to 7, since
fit_transform and the strategy's own body are now separate frames.
