+++ src/tabpfn/regressor.py
@@ -82,9 +82,10 @@ from tabpfn.preprocessing import (
 )
 from tabpfn.preprocessing.clean import clean_data_transform
 from tabpfn.preprocessing.datamodel import Feature, FeatureModality, FeatureSchema
-from tabpfn.preprocessing.date_encoding import (
+from tabpfn.preprocessing.datetimes import (
     DateTransformer,
     convert_dates,
+    make_date_transformer,
 )
 from tabpfn.preprocessing.ensemble import (
     TabPFNEnsemblePreprocessor,
@@ -878,7 +879,7 @@ class TabPFNRegressor(RegressorMixin, BaseEstimator):
         capture_input_shape(X, estimator=self, reset=True)
 
         # Must run before validation: a real datetime64 column crashes it otherwise.
-        date_transformer = DateTransformer(
+        date_transformer = make_date_transformer(
             categorical_indices=self.categorical_features_indices,
             transform_dates=self.inference_config_.TRANSFORM_DATES,
         )
